address comment move TestBatchedMessageSender to an individual file
