+++ tests/conftest.py
@@ -10,8 +10,10 @@ import time
 
 # Third Party
 import pytest
+import torch
 
 # First Party
+from lmcache.config import LMCacheEngineMetadata
 from lmcache.v1.cache_engine import LMCacheEngineBuilder
 from lmcache.v1.memory_management import MixedMemoryAllocator
 
@@ -486,3 +488,18 @@ def use_shared_allocator(request, monkeypatch, memory_allocator):
         _create_shared_allocator,
     )
     yield
+
+
+@pytest.fixture(scope="function")
+def lmcache_engine_metadata(role="worker"):
+    """Create a fresh LMCacheEngineMetadata for each test."""
+    return LMCacheEngineMetadata(
+        model_name="test_model",
+        world_size=1,
+        worker_id=0,
+        fmt="vllm",
+        kv_dtype=torch.bfloat16,
+        kv_shape=(32, 2, 256, 32, 128),
+        use_mla=False,
+        role=role,
+    )
