+++ tests/v1/storage_backend/test_local_cpu_backend.py
@@ -1,7 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 # Standard
 import threading
-import time
 
 # Third Party
 import pytest
@@ -10,6 +9,7 @@ import torch
 # First Party
 from lmcache.observability import LMCStatsMonitor
 from lmcache.utils import CacheEngineKey
+from lmcache.v1.cache_controller.message import BatchedKVOperationMsg, OpType
 from lmcache.v1.config import LMCacheEngineConfig
 from lmcache.v1.memory_management import (
     AdHocMemoryAllocator,
@@ -303,29 +303,14 @@ class TestLocalCPUBackend:
 
         local_cpu_backend.memory_allocator.close()
 
-    def test_remove_with_worker(self, memory_allocator):
+    def test_remove_with_worker(self, memory_allocator, lmcache_engine_metadata):
         """Test remove() with LMCacheWorker."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-
         config = create_test_config()
         lmcache_worker = MockLMCacheWorker()
 
-        # Create mock metadata for batched message sender initialization
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
         backend = LocalCPUBackend(
             config=config,
-            metadata=metadata,
+            metadata=lmcache_engine_metadata,
             memory_allocator=memory_allocator,
             lmcache_worker=lmcache_worker,
         )
@@ -342,10 +327,6 @@ class TestLocalCPUBackend:
         # Manually flush to ensure messages are sent for testing
         if backend.batched_msg_sender is not None:
             backend.batched_msg_sender.flush()
-
-        # Check that batched messages were sent
-        # First Party
-        from lmcache.v1.cache_controller.message import BatchedKVOperationMsg, OpType
 
         # Check that we have batched messages
         batched_msgs = [
@@ -531,486 +512,3 @@ class TestLocalCPUBackend:
         local_cpu_backend.remove(key)
         assert memory_obj.get_ref_count() == initial_ref_count + 1
         local_cpu_backend.memory_allocator.close()
-
-
-class TestBatchedMessageSender:
-    """Test cases for BatchedMessageSender."""
-
-    def teardown_method(self, method):
-        # Clean up any lingering BatchedMessageSender instances
-        # Standard
-        import gc
-
-        # First Party
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        for obj in gc.get_objects():
-            if isinstance(obj, BatchedMessageSender) and hasattr(obj, "running"):
-                obj.close()
-
-        LMCStatsMonitor.unregister_all_metrics()
-        LMCStatsMonitor.DestroyInstance()
-
-    def test_basic_batching(self):
-        """Test basic message batching functionality."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 3
-        config.extra_config["kv_msg_batch_timeout"] = 0.1
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        # Add operations below batch size threshold
-        sender.add_kv_op(OpType.ADMIT, 1)
-        sender.add_kv_op(OpType.ADMIT, 2)
-
-        # Should not have sent messages yet
-        assert len(lmcache_worker.messages) == 0
-
-        # Add one more to reach batch size
-        sender.add_kv_op(OpType.ADMIT, 3)
-
-        # Give some time for the message to be sent
-        time.sleep(0.05)
-
-        # Should have sent one batched message
-        assert len(lmcache_worker.messages) == 1
-        msg = lmcache_worker.messages[0]
-        assert len(msg.operations) == 3
-
-        sender.close()
-
-    def test_timeout_based_flush(self):
-        """Test that messages are flushed based on timeout."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 100
-        config.extra_config["kv_msg_batch_timeout"] = 0.05
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        # Add a few operations (below batch size)
-        sender.add_kv_op(OpType.ADMIT, 1)
-        sender.add_kv_op(OpType.EVICT, 2)
-
-        # Wait for timeout to trigger flush
-        time.sleep(0.1)
-
-        # Should have sent messages due to timeout
-        assert len(lmcache_worker.messages) >= 1
-
-        sender.close()
-
-    def test_manual_flush(self):
-        """Test manual flush functionality."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 100
-        config.extra_config["kv_msg_batch_timeout"] = 10.0
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        # Add operations
-        sender.add_kv_op(OpType.ADMIT, 1)
-        sender.add_kv_op(OpType.EVICT, 2)
-
-        # Manually flush
-        sender.flush()
-
-        # Should have sent messages
-        assert len(lmcache_worker.messages) == 1
-        msg = lmcache_worker.messages[0]
-        assert len(msg.operations) == 2
-
-        sender.close()
-
-    def test_sequence_numbers(self):
-        """Test that sequence numbers are monotonically increasing."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 10
-        config.extra_config["kv_msg_batch_timeout"] = 0.1
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        # Add multiple operations
-        for i in range(15):
-            sender.add_kv_op(OpType.ADMIT, i)
-
-        sender.flush()
-
-        # Collect all sequence numbers
-        seq_nums = []
-        for msg in lmcache_worker.messages:
-            for op in msg.operations:
-                seq_nums.append(op.seq_num)
-
-        # Verify sequence numbers are strictly consecutive
-        # Since we have a single consumer thread, messages must arrive in order
-        assert seq_nums == list(range(len(seq_nums))), (
-            "Sequence numbers must be consecutive starting from 0"
-        )
-
-        sender.close()
-
-    def test_concurrent_producers(self):
-        """Stress test with multiple producer threads."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 50
-        config.extra_config["kv_msg_batch_timeout"] = 0.01
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        num_threads = 10
-        ops_per_thread = 100
-
-        def producer_task(thread_id):
-            for i in range(ops_per_thread):
-                key = thread_id * ops_per_thread + i
-                sender.add_kv_op(OpType.ADMIT, key)
-
-        threads = [
-            threading.Thread(target=producer_task, args=(i,))
-            for i in range(num_threads)
-        ]
-
-        for thread in threads:
-            thread.start()
-        for thread in threads:
-            thread.join()
-
-        # Flush remaining messages
-        sender.flush()
-
-        # Collect all operations
-        all_ops = []
-        for msg in lmcache_worker.messages:
-            all_ops.extend(msg.operations)
-
-        # Verify total number of operations
-        assert len(all_ops) == num_threads * ops_per_thread
-
-        # Verify sequence numbers are strictly consecutive
-        # Even with multiple producer threads, the single consumer thread
-        # ensures messages are sent in the order they were queued.
-        seq_nums = [op.seq_num for op in all_ops]
-        assert seq_nums == list(range(num_threads * ops_per_thread)), (
-            "Sequence numbers must be consecutive starting from 0"
-        )
-
-        sender.close()
-
-    def test_mixed_operations_order(self):
-        """Test that admit and evict operations maintain order."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 100
-        config.extra_config["kv_msg_batch_timeout"] = 0.1
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        # Add operations in specific order
-        operations = [
-            (OpType.ADMIT, 1),
-            (OpType.EVICT, 1),
-            (OpType.ADMIT, 1),
-            (OpType.ADMIT, 2),
-            (OpType.EVICT, 2),
-        ]
-
-        for op_type, key in operations:
-            sender.add_kv_op(op_type, key)
-
-        sender.flush()
-
-        # Collect all operations
-        all_ops = []
-        for msg in lmcache_worker.messages:
-            all_ops.extend(msg.operations)
-
-        # Verify operations are in the same order
-        assert len(all_ops) == len(operations)
-        for i, (expected_op_type, expected_key) in enumerate(operations):
-            assert all_ops[i].op_type == expected_op_type
-            assert all_ops[i].key == expected_key
-
-        sender.close()
-
-    def test_high_throughput_stress(self):
-        """Stress test with high throughput operations."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 100
-        config.extra_config["kv_msg_batch_timeout"] = 0.01
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        num_operations = 10000
-
-        # Add many operations rapidly
-        for i in range(num_operations):
-            op_type = OpType.ADMIT if i % 2 == 0 else OpType.EVICT
-            sender.add_kv_op(op_type, i)
-
-        sender.flush()
-
-        # Collect all operations
-        all_ops = []
-        for msg in lmcache_worker.messages:
-            all_ops.extend(msg.operations)
-
-        # Verify all operations were sent
-        assert len(all_ops) == num_operations
-
-        # Verify sequence numbers are strictly increasing
-        # In a concurrent environment, we need to ensure that:
-        # 1. Sequence numbers are strictly increasing within each batch
-        # 2. No sequence numbers are duplicated
-        # 3. All sequence numbers from 0 to num_operations-1 are present
-
-        seq_nums = [op.seq_num for op in all_ops]
-
-        # Verify no duplicate sequence numbers
-        assert len(seq_nums) == len(set(seq_nums)), "Sequence numbers must be unique"
-
-        # Verify all sequence numbers are present (no messages lost)
-        expected_seq_nums = set(range(num_operations))
-        actual_seq_nums = set(seq_nums)
-        assert expected_seq_nums == actual_seq_nums, (
-            f"Missing sequence numbers: {expected_seq_nums - actual_seq_nums}"
-        )
-
-        # Verify sequence numbers are strictly increasing within each batch
-        # Since we have a single consumer thread, each batch should be processed
-        # in the order it was created, and within each batch, messages should be
-        # in the order they were added to the queue.
-        for msg in lmcache_worker.messages:
-            batch_seq_nums = [op.seq_num for op in msg.operations]
-            for i in range(1, len(batch_seq_nums)):
-                assert batch_seq_nums[i] > batch_seq_nums[i - 1], (
-                    f"Sequence numbers must be strictly increasing within batch: "
-                    f"{batch_seq_nums[i - 1]} -> {batch_seq_nums[i]}"
-                )
-
-        sender.close()
-
-    def test_close_with_pending_messages(self):
-        """Test that close() flushes pending messages."""
-        # First Party
-        from lmcache.config import LMCacheEngineMetadata
-        from lmcache.v1.cache_controller.message import OpType
-        from lmcache.v1.storage_backend.batched_message_sender import (
-            BatchedMessageSender,
-        )
-
-        config = create_test_config()
-        config.extra_config = {}
-        config.extra_config["kv_msg_batch_size"] = 100
-        config.extra_config["kv_msg_batch_timeout"] = 10.0
-
-        metadata = LMCacheEngineMetadata(
-            model_name="test_model",
-            world_size=1,
-            worker_id=0,
-            fmt="vllm",
-            kv_dtype=torch.bfloat16,
-            kv_shape=(32, 2, 256, 32, 128),
-            use_mla=False,
-            role="worker",
-        )
-
-        lmcache_worker = MockLMCacheWorker()
-        sender = BatchedMessageSender(
-            metadata=metadata,
-            config=config,
-            location="test_location",
-            lmcache_worker=lmcache_worker,
-        )
-
-        # Add operations
-        sender.add_kv_op(OpType.ADMIT, 1)
-        sender.add_kv_op(OpType.EVICT, 2)
-
-        # Close should flush pending messages
-        sender.close()
-
-        # Should have sent messages
-        assert len(lmcache_worker.messages) >= 1
-
-        # Collect all operations
-        all_ops = []
-        for msg in lmcache_worker.messages:
-            all_ops.extend(msg.operations)
-
-        assert len(all_ops) == 2
